Initial GUI Setup and icons.
Made icons for menu.

Initial Layout and general idea of how the program is going to be
structured, need to make the layout scrollable and nicer as well as add
menu items to create new 'lines' and then hook it all up to a save/open
system with a protocol to save the lyrics and their sound data.

The protocol I'm thinking is likely going to be something like PASCAL
strings, where I say something like this:

<Line Length><LINE><Sound Data Length><Sound Data> ...
<Line Length><LINE><Sound Data Length><Sound Data> ...
<Line Length><LINE><Sound Data Length><Sound Data> ...

etc ...